these might overflow on an old compiler, make it long

+++ makemono.c
@@ -132,7 +132,8 @@ is_mono(const struct rgb *c)
 static int
 color_to_mono(const struct rgb *c)
 {
-	return c->r * 299 / 1000 + c->g * 587 / 1000 + c->b * 114 / 1000;
+	return (int)(c->r * 299L / 1000 + c->g * 587L / 1000 +
+	    c->b * 114L / 1000);
 }
 
 static void
